feat: 메인 상품 목록 서버 초기 데이터 적용
