+++ src/hooks/categories/useCategories.ts
@@ -5,9 +5,14 @@ import { useQuery } from '@tanstack/react-query';
 import type { Category } from '@/types/category';
 import { categoryApi } from '@/api/categories/categoryApi';
 
-export function useCategories() {
+interface UseCategoriesOptions {
+  initialData?: Category[];
+}
+
+export function useCategories(options?: UseCategoriesOptions) {
   return useQuery<Category[]>({
     queryKey: ['categories', 'list', 'all'],
     queryFn: () => categoryApi.getCategories(),
+    initialData: options?.initialData,
   });
 }
